requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.7.1
Django==5.0.1
django-tinymce==3.6.1
idna==3.6
oauthlib==3.2.2
pillow==10.2.0
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
urllib3==2.1.0
webencodings==0.5.1
